refactorizo logica para hacer consumo de apis
